Fix iui/linux/EEKeeperWindow.ui
In order to build EEKeeper on linux it was necessary to point
ItemBrowserWidget in EEKeeperWindow.ui to proper header file, i.e.
ItemBrowserWidget.h. When editing that file QtCreator asked about
location of .qrc file (which could not be found), so that had to
be changed as well. QtCreator has also automaticaly adjusted
menuBar height by 3 pixels, so it shouldn't even be visible.